backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiosqlite>=0.20.0
pydantic>=2.0.0
httpx>=0.27.0
Pillow>=10.0.0
pyaxmlparser>=0.3.28
openai>=1.0.0
anthropic>=0.39.0
google-genai>=1.0.0
rapidocr-onnxruntime>=1.4.0
winocr>=0.0.15
tokenizers>=0.19.0
ctranslate2>=4.0.0
yt-dlp>=2024.0.0
numpy>=1.24.0
opencv-python-headless>=4.9.0
optimum[onnxruntime]>=1.17.0
transformers>=4.40.0
